requirements.txt
# Python dependencies
faker>=18.0.0
python-dotenv>=1.0.0
openai>=1.0.0
